Se creo el seeder y la factory para los productos

+++ database/migrations/2023_03_26_021538_create_productos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('marca');
             $table->string('categoria');
-            $table->integer('folio');
+            $table->bigInteger('folio');
             $table->timestamps();
         });
     }
