+++ database/migrations/2023_05_07_051822_create_existencias_table.php
@@ -15,7 +15,7 @@ class CreateExistenciasTable extends Migration
     {
         Schema::create('existencias', function (Blueprint $table) {
             $table->id();
-            $table->integer('cantidad');
+            $table->bigInteger('cantidad');
             $table->string('ubicacion');
             $table->foreignId('producto_id')->constrained();
